Simplify MappedQueryResults
