+++ packages/graph-explorer/src/connector/useGEFetchTypes.ts
@@ -4,7 +4,6 @@ import {
   VertexTypeConfig,
 } from "@/core";
 import { ConnectionConfig } from "@shared/types";
-import { MappedQueryResults } from "./gremlin/mappers/mapResults";
 import { Edge, EdgeId, Vertex, VertexId } from "@/core";
 
 export type QueryOptions = RequestInit & {
@@ -74,6 +73,31 @@ export type Criterion = {
    */
   dataType?: "String" | "Number" | "Date";
 };
+
+export type ScalarValue = number | string | Date;
+
+/**
+ * Results from any query.
+ *
+ * Used by keyword search and raw query. Could potentially be used for other
+ * queries in the future.
+ */
+export type MappedQueryResults = {
+  vertices: Vertex[];
+  edges: Edge[];
+  scalars: ScalarValue[];
+};
+
+/** Constructs a `MappedQueryResults` instance without providing all values. */
+export function toMappedQueryResults(
+  value: Partial<MappedQueryResults>
+): MappedQueryResults {
+  return {
+    vertices: value.vertices ?? [],
+    edges: value.edges ?? [],
+    scalars: value.scalars ?? [],
+  };
+}
 
 export type NeighborsRequest = {
   /**
